Updated benchmarks and tests
