+++ benchmarks/Diffusion3D.hpp
@@ -26,8 +26,8 @@ auto makeMesh(const lstr::MpiComm& comm, const auto& problem_def)
 template < CondensationPolicy CP, OperatorEvaluationStrategy S >
 void solveDiffusion3DProblem()
 {
-    std::cerr << std::format(
-        "*** Number of threads ***\nAccording to TBB: {}\nAccording to OpenMP: {}\nhwloc: {}\n\n",
+    std::println(stderr,
+        "*** Number of threads ***\nAccording to TBB: {}\nAccording to OpenMP: {}\nhwloc: {}\n",
         oneapi::tbb::global_control::active_value(oneapi::tbb::global_control::max_allowed_parallelism),
         omp_get_num_threads(),
         util::GlobalResource< util::hwloc::Topology >::getMaybeUninitialized().getNHwThreads());
